Improve registration process

+++ backend/src/main/java/io/github/k_gregory/registry/repository/CommentRepository.java
@@ -3,5 +3,5 @@ package io.github.k_gregory.registry.repository;
 import io.github.k_gregory.registry.model.Comment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Integer> {
+public interface CommentRepository extends CrudRepository<Comment, Long> {
 }
